+++ Sources/EmojiKit/Extensions/ScrollViewProxy+Grid.swift
@@ -8,9 +8,17 @@
 
 import SwiftUI
 
-extension ScrollViewProxy {
+public extension ScrollViewProxy {
 
-    /// Calculate the number of items per emoji grid row.
+    /// Scroll to a certain emoji category.
+    func scrollToCategory(
+        _ category: EmojiCategory
+    ) {
+        scrollTo(category.id)
+    }
+
+    /// Scroll to a certain emoji category, provided that it
+    /// exists in a certain category collection.
     func scrollToCategory(
         _ category: EmojiCategory?,
         in categories: [EmojiCategory]
@@ -22,13 +30,25 @@ extension ScrollViewProxy {
     }
 
     /// Calculate the number of items per emoji grid row.
+    func scrollToEmoji(
+        _ category: EmojiCategory?,
+        in categories: [EmojiCategory]
+    ) {
+        guard let category else { return }
+        let match = categories.first { $0.id == category.id }
+        guard let match, let emoji = match.emojis.first else { return }
+        scrollToSelection(.init(emoji: emoji, category: match))
+    }
+
+    /// Calculate the number of items per emoji grid row.
     func scrollToSelection(
-        _ selection: Emoji.GridSelection?
+        _ selection: Emoji.GridSelection?,
+        anchor: UnitPoint = .top
     ) {
         guard
             let category = selection?.category,
             let emoji = selection?.emoji
         else { return }
-        scrollTo(emoji.id(in: category), anchor: .top)
+        scrollTo(emoji.id(in: category), anchor: anchor)
     }
 }