+++ Sources/EmojiKit/_Deprecated/Emoji+Selection.swift
@@ -1,11 +1,3 @@
-//
-//  Emoji+Selection.swift
-//  EmojiKit
-//
-//  Created by Daniel Saidi on 2025-01-08.
-//  Copyright © 2025-2026 Daniel Saidi. All rights reserved.
-//
-
 import Foundation
 
 public extension Emoji {